BookStore/requirements.txt
SQLAlchemy==2.0.21
python-dotenv==1.0.0